aggiunge BaseProblemBuilder e modifica coerentemente ArgumentConstructor e LogicFormulaBuilder
